+++ script/lib/Deployer.sol
@@ -32,7 +32,7 @@ abstract contract Deployer is Script2 {
   ToyENS remoteEns = ToyENS(address(bytes20(hex"decaf0")));
 
   bool writeDeploy; // whether to write a .json file with updated addresses
-  address broadcaster; // who will broadcast
+  address _broadcaster; // who will broadcast
 
   constructor() {
     vm.label(address(fork), "Deployer:Fork");
@@ -64,20 +64,31 @@ abstract contract Deployer is Script2 {
   // broadcast using forge-provided tx.origin; or if default try to use <NETWORK>_PRIVATE_KEY env var's associated address.
   // In practice, this means you can set your deployer key MUMBAI_PRIVATE_KEY in your .env, and you can override that using --private-key <pk>
   function broadcast() public {
-    /* Memoize broadcaster. Cannot just do it in constructor because tx.origin for script constructors does not depend on additional CLI args */
-    if (broadcaster == address(0)) {
-      broadcaster = tx.origin;
+    vm.broadcast(broadcaster());
+  }
+
+  // compute & memoize the current broadcaster address
+  function broadcaster() public returns (address) {
+    /* Memoize _broadcaster. Cannot just do it in constructor because tx.origin for script constructors does not depend on additional CLI args */
+    if (_broadcaster == address(0)) {
+      // In the default case, forge sets the broadcaster to be tx.origin.
+      // Using msg.sender would not work since we don't know how deep in the callstack we are.
+      _broadcaster = tx.origin;
       // 0x00a3... is the default tx.origin
-      if (broadcaster == 0x00a329c0648769A73afAc7F9381E08FB43dBEA72) {
+      console.log(_broadcaster);
+      if (
+        _broadcaster == 0x00a329c0648769A73afAc7F9381E08FB43dBEA72
+          || _broadcaster == 0x1804c8AB1F12E6bbf3894d4083f33e07309d1f38
+      ) {
         string memory envVar = string.concat(simpleCapitalize(fork.NAME()), "_PRIVATE_KEY");
         try vm.envUint(envVar) returns (uint key) {
-          broadcaster = vm.rememberKey(key);
+          _broadcaster = vm.rememberKey(key);
         } catch {
           console.log("%s not found or not parseable as uint, using default broadcast sender", envVar);
         }
       }
     }
-    vm.broadcast(broadcaster);
+    return _broadcaster;
   }
 
   // buffer for output file