+++ src/MgvLib.sol
@@ -43,7 +43,7 @@ library MgvLib {
     MgvStructs.LocalPacked local;
   }
 
-  /* <a id="MgvLib/OrderResult"></a> `OrderResult` holds additional data for the maker and is given to them _after_ they fulfilled an offer. It gives them their own returned data from the previous call, and an `mgvData` specifying whether the Mangrove encountered an error. */
+  /* <a id="MgvLib/OrderResult"></a> `OrderResult` holds additional data for the maker and is given to them _after_ they fulfilled an offer. It gives them their own returned data from the previous call, and an `mgvData` specifying whether Mangrove encountered an error. */
 
   struct OrderResult {
     /* `makerdata` holds a message that was either returned by the maker or passed as revert message at the end of the trade execution*/
@@ -60,7 +60,7 @@ contract HasMgvEvents {
   event NewMgv();
 
   /* Mangrove adds or removes wei from `maker`'s account */
-  /* * Credit event occurs when an offer is removed from the Mangrove or when the `fund` function is called*/
+  /* * Credit event occurs when an offer is removed from Mangrove or when the `fund` function is called*/
   event Credit(address indexed maker, uint amount);
   /* * Debit event occurs when an offer is posted or when the `withdraw` function is called */
   event Debit(address indexed maker, uint amount);
@@ -163,7 +163,7 @@ interface IMaker {
   */
   function makerExecute(MgvLib.SingleOrder calldata order) external returns (bytes32);
 
-  /* Called after all offers of an order have been executed. Posthook of the last executed order is called first and full reentrancy into the Mangrove is enabled at this time. `order` recalls key arguments of the order that was processed and `result` recalls important information for updating the current offer. (see [above](#MgvLib/OrderResult))*/
+  /* Called after all offers of an order have been executed. Posthook of the last executed order is called first and full reentrancy into Mangrove is enabled at this time. `order` recalls key arguments of the order that was processed and `result` recalls important information for updating the current offer. (see [above](#MgvLib/OrderResult))*/
   function makerPosthook(MgvLib.SingleOrder calldata order, MgvLib.OrderResult calldata result) external;
 }